+++ packages/components/FocusZone/macos/RCTFocusZone.h
@@ -15,6 +15,6 @@ typedef NS_ENUM(NSInteger, FocusZoneDirection) {
 @property(nonatomic) BOOL disabled;
 @property(nonatomic) FocusZoneDirection focusZoneDirection;
 @property(nonatomic) NSString *navigateAtEnd;
-@property(nonatomic) NSView *defaultTabbableElement;
+@property(nonatomic) NSView *defaultKeyView;
 
 @end