+++ packages/experimental/Button/src/ToggleButton/ToggleButtonColorTokens.windows.ts
@@ -6,18 +6,16 @@ export const defaultToggleButtonColorTokens: TokenSettings<ToggleButtonTokens, T
   checked: {
     color: t.colors.neutralForeground1Selected,
     backgroundColor: t.colors.neutralBackground1Selected,
-    hovered: {
-      color: t.colors.neutralForeground1Hover,
-      backgroundColor: t.colors.neutralBackground1Hover,
+    borderColor: t.colors.neutralStroke1,
+    primary: {
+      color: t.colors.neutralForegroundOnBrandSelected,
+      backgroundColor: t.colors.brandBackgroundSelected,
+      borderColor: t.colors.brandBackgroundSelected,
     },
     subtle: {
       color: t.colors.neutralForeground1Selected,
-      backgroundColor: t.colors.neutralBackground1Selected,
-      hovered: {
-        color: t.colors.neutralForeground1Hover,
-        backgroundColor: t.colors.neutralBackground1Hover,
-        borderColor: t.colors.neutralStroke1,
-      },
+      backgroundColor: t.colors.subtleBackgroundSelected,
+      borderColor: t.colors.subtleBackgroundSelected,
     },
   },
 });