+++ packages/theming/theme-tokens/src/highContrast/tokens-alias.ts
@@ -9,7 +9,7 @@ function processAliasTokens(aliasTokens: any) {
     for (const innerKey in aliasTokens[key]) {
       const entry = aliasTokens[key][innerKey];
       if (typeof entry === 'string' && entry.includes('PlatformColor')) {
-        const color = entry.substring(14, entry.length - 1);
+        const color = 'SystemColor' + entry.substring(14, entry.length - 1) + 'Color';
         aliasTokens[key][innerKey] = PlatformColor(color);
       }
     }