+++ packages/experimental/Badge/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @fluentui-react-native/badge
 
-This log was last generated on Wed, 26 Jan 2022 20:31:04 GMT and should not be manually modified.
+This log was last generated on Wed, 02 Feb 2022 22:09:55 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.9
+
+Wed, 02 Feb 2022 22:09:55 GMT
+
+### Patches
+
+- Empty bump (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/tokens to v0.11.5
+- Bump @fluentui-react-native/use-styling to v0.8.2
+
+## 0.0.8
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-text to v0.7.12
+- Bump @fluentui-react-native/framework to v0.7.12
+- Bump @fluentui-react-native/icon to v0.11.6
+- Bump @fluentui-react-native/interactive-hooks to v0.12.2
+- Bump @fluentui-react-native/theme-tokens to v0.12.0
+- Bump @fluentui-react-native/tokens to v0.11.5
+- Bump @fluentui-react-native/use-styling to v0.8.2
+
+## 0.0.7
+
+Thu, 27 Jan 2022 10:39:35 GMT
+
+### Patches
+
+- Added Presence Badge (v.kozlova13@gmail.com)
+- Bump @fluentui-react-native/experimental-text to v0.7.11
+- Bump @fluentui-react-native/icon to v0.11.5
+- Bump @fluentui-react-native/theme-tokens to v0.11.6
+- Bump @fluentui-react-native/tokens to v0.11.5
+- Bump @fluentui-react-native/use-styling to v0.8.2
 
 ## 0.0.6